add parallelism to the new emat/eref calculators, and update the progress estimators

+++ src/edu/duke/cs/osprey/tools/Progress.java
@@ -285,7 +285,7 @@ public class Progress {
 		return incrementProgress(1);
 	}
 	
-	public boolean incrementProgress(int numWorkDone) {
+	public boolean incrementProgress(long numWorkDone) {
 		return setProgress(currentWork + numWorkDone);
 	}
